backend/requirements.txt
flask
flask-restx
psycopg2
PyJWT
flask-cors